Modularized routes and tested application
