+++ Develop/db/db.json
@@ -1,1 +1,7 @@
-[]
+[
+    {
+        "title": "Hello",
+        "text": "Bye",
+        "id": "7630"
+    }
+]
